代码优化调整

+++ cloud-common/src/main/java/cn/roamblue/cloud/common/bean/OsMemory.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class OsMemory {
-    private int memory;
+    private long memory;
 }
